Data-Engineering/Apache-Beam
Created a separate class that connects to the database, fetches vehicle data based on an argument query and returns a PCollection containing that data in the form of Vehicle objects.

Tested applying PTransforms on PCollections using ParDos and DoFns .

Changed Apache Beam dependency version to 2.36.0.

Added @ToString lombok annotation to Vehicle.